adding_foreign_keys_to_contacts_and_mails_table

+++ database/migrations/2019_07_02_083202_create_mails_table.php
@@ -15,7 +15,7 @@ class CreateMailsTable extends Migration
     {
         Schema::create('mails', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('contact_id');
+            $table->unsignedBigInteger('contact_id');
             $table->timestamps();
         });
     }
