+++ database/migrations/2019_07_26_071924_add_user_id_column_to_contacts.php
@@ -14,7 +14,7 @@ class AddUserIdColumnToContacts extends Migration
     public function up()
     {
         Schema::table('contacts', function (Blueprint $table) {
-            $table->integer('user_id')->index()->nullable();
+            $table->unsignedBigInteger('user_id')->index()->nullable();
         });
     }
